Add component with text typing effect for msg
